Return acquire result and validity in a new acquire function.
As described in [distlock doc](http://redis.io/topics/distlock): "
At this point we need to better specify our mutual exclusion rule: it
is guaranteed only as long as the client holding the lock will
terminate its work within the lock validity time (as obtained in step 3),
minus some time (just a few milliseconds in order to compensate for clock
drift between processes)." The client that acquirs a lock needs to know
the validity of lock, so it is a reasonable way to return lock validity
in the acquire function. This commit adds a new acquire function named
as acquire_with_validity to remain backwards compatibility for acquire.