Replace datetime use with a monotonic clock
A monotonic clock is one never goes backwards. It does not rely on the
system's date & time settings. It is most useful as a relative time for
calculating time deltas, not as an absolute time. Use this instead of
the datetime module for calculating elapsed times. In Python, this
feature is accessed through the time.monotonic() interface. This feature
is new to Python 3 so for Python 2.7, fallback to time.time().

https://docs.python.org/3/library/time.html#time.monotonic

> time.monotonic() → float
>
> Return the value (in fractional seconds) of a monotonic clock, i.e. a
> clock that cannot go backwards. The clock is not affected by system
> clock updates. The reference point of the returned value is undefined,
> so that only the difference between the results of consecutive calls
> is valid.